requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
alembic==1.13.1
PyMySQL==1.1.1
cryptography==45.0.4
asyncmy==0.2.10
